index.html
Main page of the Chat. When opened, it shows a login screen, with a button that opens a registration page.

After a successful login, the page then shows the chatbox with all and private chat. All private messages contain the information about the sender and the receiver, and all private messages are stored on the right box. On the left box will be displayed public (all logged users see them) messages, which contain information about their sender. If a user wants to send a private message, he can choose one of the logged users from a dropdown menu below message input field (--all-- is used for public messaging)